feat: Intro Modal 분리, 추가 구현
게임 설명 모달이 하나로는 부족한 것 같아 2가지로 분리, 설명 또한 따로 하였습니다.
