+++ src/components/main/MainContainer.jsx
@@ -1,12 +1,13 @@
 import { useState } from 'react';
 import '../../styles/main/main.css';
 import GameIntroModal from './GameIntroModal';
+import GameIntroModal2 from './GameIntroModal2';
 import GamePlayingModal from './playing/GamePlayingModal';
 import GameResultModal from './GameResultModal';
 
 export default function MainContainer() {
     // 게임 진행 상태 관리
-    // 개임 진행 상태: intro(시작 전), result(게임 종료), playing(게임 진행 중)
+    // 개임 진행 상태: intro(시작 전), intro2(게임 소개), result(게임 종료), playing(게임 진행 중)
     const [gameState, setGameState] = useState('intro');
     const [gameKey, setGameKey] = useState(0); // 게임 상태를 초기화하기 위한 키
 
@@ -18,6 +19,7 @@ export default function MainContainer() {
     return (
         <main className="main">
             {gameState === 'intro' && <GameIntroModal setGameState={setGameState} />}
+            {gameState === 'intro2' && <GameIntroModal2 setGameState={setGameState} />}
             {gameState === 'playing' && <GamePlayingModal gameState={gameState} setGameState={setGameState} />}
             {gameState === 'result' && <GameResultModal setGameState={setGameState} onRestart={handleRestart} />}
         </main>
