+++ src/components/main/playing/Answer.jsx
@@ -18,30 +18,30 @@ export default function Answer({ userInputValue, setGameState, playTime }) {
                 }
             );
         }*/
-            const postResult = async (data) => {
-                try {
-                    const token = process.env.REACT_APP_TOKEN; // 토큰 가져오기
-                    const res = await fetch(
-                        'https://0by7j8suf2.execute-api.ap-northeast-2.amazonaws.com/proxy/api/result',
-                        {
-                            method: 'POST',
-                            mode: 'cors',
-                            headers: {
-                                'Content-Type': 'application/json',
-                                'Authorization': `Bearer ${token}`, // Authorization 헤더에 토큰 삽입
-                            },
-                            body: data, // 서버로 보낼 데이터
-                        }
-                    );
-    
-                    if (!res.ok) throw new Error(`HTTP ${res.status}`);
-    
-                    console.log('점수 전송 성공');
-                } catch (e) {
-                    console.error('전송 오류:', e);
-                }
-            };
-        
+        const postResult = async (data) => {
+            try {
+                const token = process.env.REACT_APP_TOKEN; // 토큰 가져오기
+                const res = await fetch(
+                    'https://0by7j8suf2.execute-api.ap-northeast-2.amazonaws.com/proxy/api/result',
+                    {
+                        method: 'POST',
+                        mode: 'cors',
+                        headers: {
+                            'Content-Type': 'application/json',
+                            'Authorization': `Bearer ${token}`, // Authorization 헤더에 토큰 삽입
+                        },
+                        body: data, // 서버로 보낼 데이터
+                    }
+                );
+
+                if (!res.ok) throw new Error(`HTTP ${res.status}`);
+
+                console.log('점수 전송 성공');
+            } catch (e) {
+                console.error('전송 오류:', e);
+            }
+        };
+
 
         if (userInputValue.length === 6) {
             const userAnswer = userInputValue.join("");
@@ -59,7 +59,7 @@ export default function Answer({ userInputValue, setGameState, playTime }) {
                     }
                     return record;
                 });
-console.log(currentStudentId);
+                console.log(currentStudentId);
                 const data = {
                     gameName: "keyzzle",
                     userId: currentStudentId,
