+++ src/styles/main/intro1.css
@@ -1,0 +1,89 @@
+/*********** 게임 소개 모달1 ***********/
+.game-intro1-container {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: stretch;
+    align-items: center;
+    margin: 0;
+    box-sizing: border-box;
+    height: 100%;
+    color: white;
+}
+
+.game-intro1-description {
+    width: 100%;
+    height: 70%;
+    text-align: center;
+}
+
+.game-intro1-description h1 {
+    font-size: 40px;
+    font-weight: bold;
+    margin-top: 20px;
+    margin-bottom: 40px;
+}
+
+.game-intro1-description h1:last-of-type {
+    margin-top: 40px;
+    margin-bottom: 20px;
+}
+
+.game-intro1-description p {
+    font-size: 25px;
+    margin-top: 20px;
+    line-height: 35px;
+}
+
+.highlight-yellow {
+    color: #FFE5B4;
+    font-weight: bold;
+    text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
+}
+
+.highlight-lightskyblue {
+    color: #87CEFA;
+    font-weight: bold;
+    text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
+}
+
+.highlight-red {
+    font-size: 25px;
+    font-weight: bold;
+    color: #FFC0CB;
+    margin-bottom: 20px;
+    animation: pulse 2s infinite;
+}
+
+@keyframes pulse {
+    0% {
+        opacity: 1;
+    }
+
+    50% {
+        opacity: 0.7;
+    }
+
+    100% {
+        opacity: 1;
+    }
+}
+
+.next-bt {
+    width: 30%;
+    height: 8%;
+    border: none;
+    border-radius: 10px;
+    background-color: rgb(12, 148, 12);
+    color: white;
+    font-size: 25px;
+    font-weight: bold;
+    cursor: pointer;
+    margin-top: 20%;
+    transition: all 0.3s ease;
+}
+
+.next-bt:hover {
+    background-color: rgb(9, 107, 9);
+    transform: scale(1.05);
+}
