+++ src/styles/main/intro2.css
@@ -1,4 +1,4 @@
-/*********** 게임 소개 모달 ***********/
+/*********** 게임 소개 모달2 ***********/
 .game-intro-container {
     width: 100%;
     display: flex;
@@ -11,30 +11,57 @@
     color: white;
 }
 
-.game-intro-description {
-    width: 100%;
-    height: 70%;
-    text-align: center;
-}
-
-.game-intro-description h1 {
+.game-intro-container h1 {
     font-size: 40px;
     font-weight: bold;
     margin-top: 20px;
     margin-bottom: 40px;
 }
 
-.game-intro-description p {
+.game-intro-container h1:last-of-type {
+    margin-top: 40px;
+    margin-bottom: 40px;
+}
+
+.game-intro-operation {
+    width: 100%;
+    height: 70%;
+    text-align: center;
+}
+
+.game-intro-operation p {
     font-size: 25px;
     margin-top: 20px;
     line-height: 35px;
 }
 
 /* 강조 텍스트 스타일 */
-.game-intro-description p .highlight {
+.game-intro-operation p .highlight-yellow {
     color: #FFD700;
     font-weight: bold;
     text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
+    animation: pulse 2s infinite;
+}
+
+.game-intro-operation p .highlight-green {
+    color: #1ad31a;
+    font-weight: bold;
+    text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
+    animation: pulse 2s infinite;
+}
+
+.game-intro-operation p .highlight-lightskyblue {
+    color: lightskyblue;
+    font-weight: bold;
+    text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.1);
+    animation: pulse 2s infinite;
+}
+
+.highlight-red {
+    font-size: 25px;
+    font-weight: bold;
+    color: #FFC0CB;
+    margin-bottom: 20px;
     animation: pulse 2s infinite;
 }
 
@@ -52,32 +79,6 @@
     }
 }
 
-/* .game-intro-description p:first-of-type {
-    font-size: 25px;
-    margin-top: 8%;
-}
-
-.game-intro-description p:last-of-type {
-    font-size: 18px;
-    margin-top: 10%;
-} */
-
-.player-id {
-    text-align: center;
-    width: 50%;
-    height: 7%;
-    margin-top: 3%;
-    margin-bottom: 3%;
-    font-size: 17px;
-    font-weight: bold;
-}
-
-.game-intro-description-highlight {
-    font-size: 20px;
-    font-weight: bold;
-    color: #FFC0CB;
-}
-
 .game-intro-input-container {
     width: 100%;
     height: 10%;
@@ -85,9 +86,10 @@
     justify-content: center;
     align-items: center;
     gap: 20px;
+    margin-top: 20px;
 }
 
-.game-intro-input-description {
+.game-intro-input-unique-id {
     font-size: 30px;
     font-weight: bold;
     color: white;
@@ -128,4 +130,12 @@
     &:hover {
         background-color: rgb(9, 107, 9);
     }
+}
+
+.game-example-img {
+    width: 60%;
+    height: auto;
+    margin: 10px 0;
+    border-radius: 20px;
+    box-shadow: 0 4px 15px rgba(0, 0, 0, 0.2);
 }
