+++ src/styles/main/playing/Submit.css
@@ -10,7 +10,7 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    gap: 5px;
+    gap: 15px;
     margin-top: 30px;
     margin-bottom: 30px;
 }
